fix issues

+++ src/main/java/com/tutor/persistance/entity/BaseEntity.java
@@ -1,21 +1,18 @@
 package com.tutor.persistance.entity;
 
 import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.MappedSuperclass;
 import lombok.Data;
+import lombok.experimental.SuperBuilder;
 
 import java.io.Serializable;
-import java.time.LocalDateTime;
 
 @Data
 @MappedSuperclass
 public abstract class BaseEntity extends AuditableEntity implements Serializable {
 
     @Column(name = "status")
-    private Long status;
+    private Integer status;
 
     @Column(name = "is_deleted")
     private Boolean isDeleted;
